% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_pairs.R
\name{count_pairs}
\alias{count_pairs}
\title{Count Pairs}
\usage{
count_pairs(n, m)
}
\arguments{
\item{n}{A numeric vector.}

\item{m}{A numeric vector.}
}
\value{
An integer count of pairs that satisfy the condition.
}
\description{
An function that takes two integers, n and m, and returns the count of integer pairs (x,y) in {1,2,...,n} such that xy mod (x+y) = m.
}
\examples{
count_pairs(1,1)
}
