% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_logit.R
\name{gen_logit}
\alias{gen_logit}
\title{Gen logit}
\usage{
gen_logit(matrix_1, x)
}
\arguments{
\item{matrix_1}{A matrix vector.}

\item{x}{A numeric vector.}
}
\value{
An probabilities for all levels of Y.
}
\description{
Calculates the probabilities for each level of Y in a generalized logit model. In a generalized logit model (multinomial logistic regression), the outcome Y is a categorical variable with K+1 levels: 0,1,...,K, where 0 is the reference level. The model relates Y to covariates X as: ln(Pr(Y+K)/Pr(Y=0)) = (β_k)^T * X. It returns all probabilities on all levels of Y.
}
\examples{
gen_logit ( matrix (0 , 2 , 2) , c (1 , 0))
}
