% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_kendall.R
\name{matrix_kendall}
\alias{matrix_kendall}
\title{Matrix kendall}
\usage{
matrix_kendall(X)
}
\arguments{
\item{X}{A matrix vector.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
An integer that the maximum across all rows.
}
\description{
This function computes Kendall's tao between each row of an m*n matrix X and y(1, ... ,n), returning the maxmium tao across all rows.
}
\examples{
matrix_kendall ( matrix ( c (1 , 2 , 2 , 1) , 2 , 2))
}
