% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_str.R
\name{num_str}
\alias{num_str}
\title{Num Str}
\usage{
num_str(input_num)
}
\arguments{
\item{input_num}{A series number in string type.}
}
\value{
An integer of the first position that the input value appears in the sequence S.
}
\description{
Defines S as an infinite string string formed by concatenating all the natural numbers in increasing order, i.e. S = "123456....99100101..."
This function that takes a digit sequence s and returns its first position in S.
}
\examples{
num_str ("12")
}
