% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic.R
\name{quadratic}
\alias{quadratic}
\title{Quadratic}
\usage{
quadratic(a, b, c)
}
\arguments{
\item{a}{A numeric vector.}

\item{b}{A numeric vector.}

\item{c}{A numeric vector.}
}
\value{
An integer of valid solution for the equation ax^2 + bx +c = 0.
}
\description{
An R function that takes three numbers a, b, and c , as arguments and returns the smallest solution of the equation ax^2 + bx + c = 0, guaranteed to have at least one solution.
}
\examples{
quadratic (1 , 4 , 1)
}
