% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{triangle}
\alias{triangle}
\title{Find triangle}
\usage{
triangle(edges)
}
\arguments{
\item{edges}{A matrix vector.}
}
\value{
An bool value that whether the input matrix has an triangle with length 3.
}
\description{
An R function, triangle, takes and K*2 matrix edges (representing and undirected graph's edges) and checks if the graph contains any triangles (cycles of length 3).
}
\examples{
triangle ( matrix ( c (1 , 2 , 2 , 3) , 2 , 2))
}
