% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_logit.R
\name{gen_logit}
\alias{gen_logit}
\title{Gen logit}
\usage{
gen_logit(matrix_1, x)
}
\arguments{
\item{matrix_1}{A matrix vector.}

\item{x}{A numeric vector.}
}
\value{
An probabilities for all levels of Y.
}
\description{
An R function that takes a K*p matrix beta and a length_p vector x, and returns the probalilities for all levels of Y.
}
\examples{
gen_logit ( matrix (0 , 2 , 2) , c (1 , 0))
}
