% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_kendall.R
\name{matrix_kendall}
\alias{matrix_kendall}
\title{Matrix kendall}
\usage{
matrix_kendall(X)
}
\arguments{
\item{X}{A matrix vector.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
An integer that the maximum τ across all rows.
}
\description{
Given a set of observations x=(x1,x2,...,xn) and y=(y1,y2,...,yn), Kendall's τ, a non-parametric rank correlation coefficient, is defined as:τ = ((number of concordant pairs) - (number of disconcordant pairs))/((n(-1))/2)
For 1≤i<j≤n, a pair of observations (xi, yi) and (xj, yj) is:
Concordant if both xi>xj, yi>yj, or if both xi<xj and yi<yj.
Discordant if xi>xj and yi<yj or ifxi\if{html}{\out{<xj and yi>}}yj.
The function should calculate Kendall's τ between each row of X and y=(1,2,...,n), then return the maximun τ value achieved across all rows of X.
}
\examples{
matrix_kendall ( matrix ( c (1 , 2 , 2 , 1) , 2 , 2))
}
